Extract attached metadata and add a test
